Move ServerCertsKeys, cert_chain_hash and read_certs_and_keys out of huginn-proxy-lib::tls::cert_source
