+++ huginn-proxy-lib/Cargo.toml
@@ -11,6 +11,7 @@ arc-swap.workspace = true
 bytes.workspace = true
 http.workspace = true
 http-body-util.workspace = true
+huginn-certs.workspace = true
 huginn-net-http.workspace = true
 huginn-net-tcp.workspace = true
 huginn-net-tls.workspace = true
