+++ huginn-proxy-lib/src/error.rs
@@ -20,5 +20,13 @@ pub enum ProxyError {
     Http(String),
 }
 
+impl From<huginn_certs::CertError> for ProxyError {
+    fn from(e: huginn_certs::CertError) -> Self {
+        match e {
+            huginn_certs::CertError::Tls(msg) => ProxyError::Tls(msg),
+        }
+    }
+}
+
 pub type ProxyResult<T> = std::result::Result<T, ProxyError>;
 pub type Result<T> = ProxyResult<T>;
