+++ huginn-proxy-lib/src/tls/mod.rs
@@ -1,6 +1,5 @@
 pub mod acceptor;
 pub mod cert_resolver;
-pub mod cert_source;
 pub mod cipher_suites;
 pub mod curves;
 pub mod metrics;
@@ -8,8 +7,8 @@ pub mod session_resumption;
 pub mod setup;
 pub use acceptor::build_server_config_with_resolver;
 pub use cert_resolver::{CertReloadReport, DynamicCertResolver};
-pub use cert_source::{cert_chain_hash, ServerCertsKeys};
 pub use cipher_suites::{is_cipher_suite_supported, supported_cipher_suites};
 pub use curves::{is_curve_supported, supported_curves};
+pub use huginn_certs::{cert_chain_hash, ServerCertsKeys};
 pub use metrics::{extract_tls_info, record_tls_handshake_metrics};
 pub use setup::build_tls_acceptor;
